webadmin/requirements.txt
# Phishly WebAdmin - Flask Application Requirements

# ============================================
# Flask Web Framework
# ============================================
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# ============================================
# Authentication & Sessions
# ============================================
Flask-Session==0.8.0      # Redis-backed sessions (updated for Werkzeug 3.x compatibility)
Flask-WTF==1.2.1          # CSRF protection and forms
Flask-Login==0.6.3        # User authentication

# ============================================
# Database
# ============================================
Flask-SQLAlchemy==3.1.1   # Flask integration for SQLAlchemy
SQLAlchemy==2.0.23        # ORM
psycopg2-binary==2.9.9    # PostgreSQL adapter
alembic==1.13.1           # Database migrations

# ============================================
# Redis Client
# ============================================
redis==5.0.1              # Session storage

# ============================================
# Task Queue (for triggering Celery tasks)
# ============================================
celery==5.3.4             # Celery client for triggering email tasks

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0

# ============================================
# Production Server
# ============================================
# gunicorn==21.2.0        # WSGI server for production